+++ android/app/src/main/java/com/bigcorps/driveraimvp/OfferAdmissionGate030.kt
@@ -95,6 +95,7 @@ object OfferAdmissionGate030 {
         Reader2Shadow030.resetRuntime()
         Reader2Parallel031.resetRuntime()
         Reader2Accumulator032.resetRuntime()
+        Reader2Consensus0321.resetRuntime()
     }
 
     fun toJson(context: Context): JSONObject {